refactor: extract CPU-clean service interface
